Move tileset error message formatting into a separate file
Allow referencing error formatting from TilesetLoader and other future source code files without creating circular include references
