+++ OP2Utility.vcxproj
@@ -118,6 +118,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="src\Sprite\TilesetCommon.cpp" />
     <ClCompile Include="src\Sprite\TilesetLoader.cpp" />
     <ClInclude Include="src\Archive\CompressionType.h" />
     <ClInclude Include="src\Archive\WaveFile.h" />
@@ -183,6 +184,7 @@
     <ClInclude Include="src\Sprite\SpriteLoader.h" />
     <ClInclude Include="src\Sprite\PaletteHeader.h" />
     <ClInclude Include="src\Sprite\SectionHeader.h" />
+    <ClInclude Include="src\Sprite\TilesetCommon.h" />
     <ClInclude Include="src\Stream\DynamicMemoryWriter.h" />
     <ClInclude Include="src\Stream\SliceReader.h" />
     <ClInclude Include="src\Stream\FileReader.h" />
