+++ OP2Utility.vcxproj.filters
@@ -183,6 +183,9 @@
     <ClInclude Include="src\Sprite\TilesetLoader.h">
       <Filter>Sprite</Filter>
     </ClInclude>
+    <ClInclude Include="src\Sprite\TilesetCommon.h">
+      <Filter>Sprite</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\XFile.cpp">
@@ -288,6 +291,9 @@
       <Filter>Sprite</Filter>
     </ClCompile>
     <ClCompile Include="src\Sprite\TilesetLoader.cpp">
+      <Filter>Sprite</Filter>
+    </ClCompile>
+    <ClCompile Include="src\Sprite\TilesetCommon.cpp">
       <Filter>Sprite</Filter>
     </ClCompile>
   </ItemGroup>
